Settings for UI (#252)
* Removed unused __future__ import for python3

* reading signed ECS values

* fix device instance

* PageBattery.qml updates for VenusOS 2.91

* silent the install script copy

* Add space for 24 cells

* activate custom name

* update qml installer to only backup qml once and add Add Battery setup page

* publish default settings

* Fixes for when battery does not have get_max_cell_voltage

* driver version moved to firmware version

* Cell voltages smaller display

* settings location update

+++ etc/dbus-serialbattery/installqml.sh
@@ -1,10 +1,13 @@
 #!/bin/sh
-#backup old PageBattery.qml
-cp /opt/victronenergy/gui/qml/PageBattery.qml /opt/victronenergy/gui/qml/PageBattery.qml.backup
+#backup old PageBattery.qml once. New firmware upgrade will remove the backup
+if [ ! -f /opt/victronenergy/gui/qml/PageBattery.qml.backup ]; then
+    cp /opt/victronenergy/gui/qml/PageBattery.qml /opt/victronenergy/gui/qml/PageBattery.qml.backup
+fi
 #copy new PageBattery.qml
 cp /data/etc/dbus-serialbattery/qml/PageBattery.qml /opt/victronenergy/gui/qml/
 #copy new PageBatteryCellVoltages
 cp /data/etc/dbus-serialbattery/qml/PageBatteryCellVoltages.qml /opt/victronenergy/gui/qml/
+cp /data/etc/dbus-serialbattery/qml/PageBatterySetup.qml /opt/victronenergy/gui/qml/
 #stop gui
 svc -d /service/gui
 #sleep 1 sec
